drivers: display: add waveshare dsi2dpi display bridge support
Replace the waveshare_dsi_lcd panel with the waveshare_dsi2dpi
display bridge.
