--- drivers/display/Kconfig.waveshare_dsi_lcd
@@ -1,12 +1,0 @@
-# Copyright 2025 NXP
-#
-# SPDX-License-Identifier: Apache-2.0
-
-config WAVESHARE_7INCH_DSI_LCD_C
-	bool "waveshare 7inch DSI LCD(C) driver"
-	default y
-	select MIPI_DSI
-	select I2C
-	depends on DT_HAS_WAVESHARE_7INCH_DSI_LCD_C_ENABLED
-	help
-	  Enable driver for waveshare 7inch DSI LCD(C).
