+++ dts/bindings/display/waveshare,dsi2dpi.yaml
@@ -4,11 +4,19 @@
 # SPDX-License-Identifier: Apache-2.0
 #
 
-description: Waveshare 7INCH DSI LCD(C)
+title: Waveshare DSI-to-DPI display bridge
 
-compatible: "waveshare,7inch-dsi-lcd-c"
+description: |
+    Waveshare display shields consist of a DPI panel and a controller board.
+    The controller board consists of ICN6211 DSI-to-RGB bridge and an MCU, to
+    convert DSI to DPI and control the backlight.
+    This binding treats the controller board as a whole unit.
+
+compatible: "waveshare,dsi2dpi"
 
 include: base.yaml
+
+on-bus: i2c
 
 properties:
   reg:
